Move play_fixed from manual.py to board.py
